eventqueue: block in `WaitToBeDrained` for in-flight events to be completed
Once an EventQueue has been closed, that means that no more events can be
Enqueued onto it. However, inflight events may still be running. Add a means of
signaling that an Event has been completed via a WaitGroup, which
`WaitToBeDrained` waits on after the queue has been closed. Once an Event has
finished, it releases its wait on this WaitGroup. There are guaranteed to be no
more Events which will be ran due to the queue being closed.

In the context of the Endpoint, which uses the EventQueue, this is useful
because we want to make sure that no more Events are pushed onto its EventQueue,
and that no more Events are being processed / are in flight that were pushed
onto its EventQueue, because during Endpoint deletion, we want to make sure that
no Events are running on its EventQueue before we clean up its maps, directories
on the filesystem, etc., which can be used while it is regenerating.

Signed-off by: Ian Vernon <ian@cilium.io>